refactor -> Exception관련 클래스 exception 패키지로 묶어서 관리
