+++ src/main/java/baseball/Application.java
@@ -1,10 +1,10 @@
 package baseball;
 
 import camp.nextstep.edu.missionutils.Randoms;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
+import Exception.*;
 
 public class Application {
