+++ src/main/java/baseball/Game.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.Scanner;
+import Exception.CheckException;
 
 
 public class Game {
